Move ABACUS_MAP and map_params function to abacus.py for better organization and clarity.
