+++ acm/hod/base.py
@@ -7,54 +7,10 @@ from abacusnbody.hod.abacus_hod import AbacusHOD
 from cosmoprimo import Cosmology
 from cosmoprimo.fiducial import DESI, AbacusSummit
 
-ABACUS_MAP = {
-    "logM1": ["log_1"],
-    "Acent": ["A_cen"],
-    "Asat": ["A_sat"],
-    "Bcent": ["B_cen"],
-    "Bsat": ["B_sat"],
-}
+from acm.utils.abacus import map_params
 
 SIM_PARAMS_MAP = {"z_mock": ["redshift"]}
 
-
-def map_params(
-    params: dict | list[str], mapping: dict[str, list[str]] = ABACUS_MAP
-) -> dict | list[str]:
-    """
-    Map custom parameters names to fixed parameters.
-
-    Parameters
-    ----------
-    params : dict | list[str]
-        Dictionary or list of custom parameters.
-    mapping : dict[str, list[str]]
-        Mapping from custom parameter names to fixed parameter names.
-        Keys are fixed parameter names, values are lists of custom parameter names that map to the fixed parameter name.
-
-    Returns
-    -------
-    dict | list[str]
-        Dictionary or list of fixed parameters. Use the same type as the input params.
-
-    Raises
-    ------
-    ValueError
-        If the type of params is not dict or list.
-    """
-    is_dict = type(params) is dict
-    if type(params) not in [dict, list]:
-        raise ValueError("Invalid type for params. Must be either dict or list.")
-
-    for abacus_key, custom_keys in mapping.items():
-        for custom_key in custom_keys:
-            if custom_key in params:  # Check if the custom key is used
-                # Replace custom key with Abacus key
-                if is_dict:
-                    params[abacus_key] = params.pop(custom_key)
-                else:  # is list
-                    params[params.index(custom_key)] = abacus_key
-    return params
 
 
 def get_abacus_simname(sim_type: str, cosmo_idx: int, phase_idx: int) -> str:
